==> 03_UploadVisualization.ipynb <==
# Save app to HISE

Save app back to HISE study space

Ensure python version is 3.9 or above for sklearn compatibility with pip-tools

## Set Up

!python --version

import sys 
import os
sys.path.insert(0, '/home/jupyter/hisepy/build/lib') # Note: reinstalled hisepy into python 3.9 conda environment. adding to path here
import hisepy as hp
import session_info
import glob

## Variables

# Destination Study space
spaceid = '40df6403-29f0-4b45-ab7d-f46d420c422e'

# Source DEG dataset
input_fileid = ['fc83b89f-fd26-43b8-ac91-29c539703a45'] 

## Data

# Data files
all_data = ['data/all_mast_deg_2023-09-06.pkl', "data/custom_gs_dict.pkl", "data/meta_color_dict.pkl","data/app_info.txt", "assets/styles.css"]
print(all_data) 
len(all_data)

sizes = [os.stat(x).st_size for x in all_data]
print("{} : {}GB".format("Total Size", round(sum(sizes)/(1024*1024*1024),3)))

for fp in all_data:
    print(fp + " exists: " + str(os.path.exists(fp)))

## Save App

hp.save_dash_app(
    app_filepath = 'app.py',
    additional_files = all_data,
    input_file_ids= input_fileid,
    study_space_id = spaceid,
    title = "In-Vitro T-cell VRd DEG Explorer, v1.0.2",
    description = "Explore DEGs for various in-vitro drug effect comparisons",
    image = "/home/jupyter/repro-vrd-tea-seq-deg-app/thumbnail.png"
)

import session_info
session_info.show()
